requirements.txt
fastapi
uvicorn
pydantic
pyjwt
psycopg2-binary
requests

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
httpx>=0.24.0
faker>=18.0.0